# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "BareFoot Nomad is application that Make company global travel and accommodation easy and convenient for the strong workforce of savvy members of staff, by leveraging the modern web.",
    "version": " ",
    "title": "BareFootNomad",
    "termsOfService": "https://github.com/Stackup-Rwanda/ateam-bn-backend/blob/develop/README.md",
    "contact": {
      "email": "the-a-team@gmail.com"
    }
  },
  "host": "ateam-bn-backend-staging.herokuapp.com",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "none"
  ],
  "produces": [
    "application/json",
    "none"
  ],
  "paths": {
...
    "/accommodation/{id}/edit": {
      "patch": {
        "description": "Travel Admin can edit an accommodation",
        "summary": "Edit Accommodation",
        "tags": [
          "Accommodation"
        ],
        "operationId": "EditAccommodation",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "security token"
          },
          {
            "in": "path",
            "name": "id",
            "type": "integer",
            "required": true,
            "description": "parsing an accommodation Id"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string",
            "description": "the name of the accommodation"
          },
          {
            "in": "formData",
            "name": "description",
            "type": "string",
            "description": "the description of the accommodation"
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "description": "upload a picture"
          },
          {
            "in": "formData",
            "name": "locationId",
            "type": "string",
            "description": "the location of the accommodation"
          },
          {
            "in": "formData",
            "name": "geoLocation",
            "type": "string",
            "description": "the geoLocation of the accommodation"
          },
          {
            "in": "formData",
            "name": "highlights",
            "type": "string",
            "description": "the highlights that happened"
          },
          {
            "in": "formData",
            "name": "amenities",
            "type": "string",
            "description": "provide the amenities the accommodation provides"
          }
        ],
        "responses": {
          "201": {
            "description": "Accommodation Successfully Updated",
            "headers": {}
          },
          "400": {
            "description": "Please select one or more pictures / token is malformed/invalid",
            "headers": {}
          },
          "415": {
            "description": "Please select the right type of image",
            "headers": {}
          },
          "401": {
            "description": "You are not allowed to access this route",
            "headers": {}
          }
        }
      }
    },
...
    "/room": {
      "post": {
        "description": "Manager can create a room",
        "summary": "Create new room",
        "tags": [
          "Room"
        ],
        "operationId": "CreateRoom",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "manager's token"
          },
          {
            "in": "formData",
            "name": "accommodationId",
            "type": "integer",
            "description": "the accommodation Id"
          },
          {
            "in": "formData",
            "name": "roomType",
            "type": "string",
            "description": "the type of the room"
          },
          {
            "in": "formData",
            "name": "amenities",
            "type": "string",
            "items": {
              "type": "string"
            },
            "description": "Amenities or facilities for the room"
          },
          {
            "in": "formData",
            "name": "cost",
            "type": "string",
            "description": "the cost of the room"
          },
          {
            "in": "formData",
            "name": "image",
            "type": "file",
            "description": "the image of this room"
          },
          {
            "in": "formData",
            "name": "status",
            "type": "string",
            "description": "The status of the room i.e booked/available"
          }
        ],
        "responses": {
          "201": {
            "description": "The room is successfully created",
            "headers": {}
          },
          "400": {
            "description": "provide the correct accommodation",
            "headers": {}
          },
          "415": {
            "description": "Please select the right type of image",
            "headers": {}
          },
          "403": {
            "description": "You are not allowed",
            "headers": {}
          }
        }
      },
      "get": {
        "description": "Retrieve all rooms",
        "summary": "View all rooms available",
        "tags": [
          "Room"
        ],
        "operationId": "ViewRooms",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "User's token"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all Rooms",
            "headers": {}
          },
          "400": {
            "description": "Malformed/ Incorrect security token ! Check token and try again.",
            "headers": {}
          }
        }
      }
    },
...
  },
...
}